+++ lib/net.dart
@@ -1,6 +1,6 @@
 /// Internet related validators.
 ///
-/// E.g.: email, ipv4, ipv6.
+/// E.g. email, ipv4, ipv6, mac, url.
 library net;
 
 export 'src/net/email.dart';